+++ src/boomerang/ssl/statements/CaseStatement.cpp
@@ -26,26 +26,42 @@ CaseStatement::CaseStatement()
 }
 
 
+CaseStatement::CaseStatement(const CaseStatement &other)
+    : GotoStatement(other)
+    , m_switchInfo(new SwitchInfo(*other.m_switchInfo))
+{
+}
+
+
 CaseStatement::~CaseStatement()
 {
 }
 
 
+CaseStatement &CaseStatement::operator=(const CaseStatement &other)
+{
+    GotoStatement::operator=(other);
+
+    m_switchInfo.reset(new SwitchInfo(*other.m_switchInfo));
+    return *this;
+}
+
+
 SwitchInfo *CaseStatement::getSwitchInfo()
 {
-    return m_switchInfo;
+    return m_switchInfo.get();
 }
 
 
 const SwitchInfo *CaseStatement::getSwitchInfo() const
 {
-    return m_switchInfo;
+    return m_switchInfo.get();
 }
 
 
-void CaseStatement::setSwitchInfo(SwitchInfo *psi)
+void CaseStatement::setSwitchInfo(std::unique_ptr<SwitchInfo> psi)
 {
-    m_switchInfo = psi;
+    m_switchInfo = std::move(psi);
 }
 
 
@@ -99,7 +115,7 @@ Statement *CaseStatement::clone() const
     ret->m_isComputed = m_isComputed;
 
     if (m_switchInfo) {
-        ret->m_switchInfo            = new SwitchInfo;
+        ret->m_switchInfo.reset(new SwitchInfo);
         *ret->m_switchInfo           = *m_switchInfo;
         ret->m_switchInfo->switchExp = m_switchInfo->switchExp->clone();
     }