+++ src/boomerang/ssl/statements/CaseStatement.h
@@ -46,12 +46,12 @@ class BOOMERANG_API CaseStatement : public GotoStatement
 {
 public:
     CaseStatement();
-    CaseStatement(const CaseStatement &other) = default;
-    CaseStatement(CaseStatement &&other)      = default;
+    CaseStatement(const CaseStatement &other);
+    CaseStatement(CaseStatement &&other) = default;
 
     virtual ~CaseStatement() override;
 
-    CaseStatement &operator=(const CaseStatement &other) = default;
+    CaseStatement &operator=(const CaseStatement &other);
     CaseStatement &operator=(CaseStatement &&other) = default;
 
 public:
@@ -86,8 +86,8 @@ public:
     SwitchInfo *getSwitchInfo();
     const SwitchInfo *getSwitchInfo() const;
 
-    void setSwitchInfo(SwitchInfo *psi);
+    void setSwitchInfo(std::unique_ptr<SwitchInfo> psi);
 
 private:
-    SwitchInfo *m_switchInfo; ///< Ptr to struct with information about the switch
+    std::unique_ptr<SwitchInfo> m_switchInfo; ///< Ptr to struct with information about the switch
 };